+++ src/main/java/roidrole/thaumicinfo/ThaumicInformation.java
@@ -1,13 +1,10 @@
 package roidrole.thaumicinfo;
 
-import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.fml.common.Loader;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.common.event.FMLLoadCompleteEvent;
 import net.minecraftforge.fml.common.event.FMLPreInitializationEvent;
-import net.minecraftforge.fml.relauncher.Side;
 import org.apache.logging.log4j.Logger;
-import roidrole.thaumicinfo.jei.ResearchManager;
 import roidrole.thaumicinfo.visualores.PacketHandler;
 
 
@@ -30,7 +27,7 @@ public class ThaumicInformation {
     @Mod.EventHandler
     public void preInit(FMLPreInitializationEvent event) {
         LOGGER = event.getModLog();
-        if(ThaumicInformationConfig.visualOresConfig.dioptraUpdatesAura && Loader.isModLoaded("visualores")){
+        if(ThaumicInformationConfig.visualOresConfig.dioptra.enabled && Loader.isModLoaded("visualores")){
             PacketHandler.preInit();
         }
     }